Deduplicate top level lighting CLI command groups (#15761)
* unify remove and delete command groups & the add and delete command groups
* added changelog
* fix tests
* Apply suggestions from code review

(cherry picked from commit 7b2788e39c113f03abaad37161653ce7f24941ea)
