+++ src/lightning_app/cli/lightning_cli.py
@@ -24,11 +24,9 @@ from lightning_app.cli.commands.connection import (
     disconnect,
 )
 from lightning_app.cli.commands.logs import logs
-from lightning_app.cli.lightning_cli_add import cli_add
 from lightning_app.cli.lightning_cli_create import create
 from lightning_app.cli.lightning_cli_delete import delete
 from lightning_app.cli.lightning_cli_list import get_list
-from lightning_app.cli.lightning_cli_remove import cli_remove
 from lightning_app.core.constants import DEBUG, ENABLE_APP_COMMENT_COMMAND_EXECUTION, get_lightning_cloud_url
 from lightning_app.runners.runtime import dispatch
 from lightning_app.runners.runtime_type import RuntimeType
@@ -381,8 +379,6 @@ def stop() -> None:
 _main.add_command(get_list)
 _main.add_command(delete)
 _main.add_command(create)
-_main.add_command(cli_add)
-_main.add_command(cli_remove)
 
 
 @_main.command("ssh")
